% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarScore.R
\name{calc_I22_mean}
\alias{calc_I22_mean}
\title{The I22 of Mean-shift perturbation model}
\usage{
calc_I22_mean(z, k, p, n, sigma_inv)
}
\arguments{
\item{z}{The matrix of independent variables.}

\item{k}{The dimension of VAR.}

\item{p}{Lags.}

\item{n}{The number of samples.}

\item{sigma_inv}{The inverse matrix of Σ.}
}
\value{
The I22 of Mean-shift perturbation model
}
\description{
The I22 of Mean-shift perturbation model
}
\examples{
library(VarScore)
x<-gene_indep_variable(IBM_SP, 1, 2)
V<-myVAR(IBM_SP)
sigma<-V$hat_sigma
B<-V$hat_B
sigma_inv<-solve(sigma)
k<-dim(sigma)[1]
p<-(dim(B)[2]-1)\%/\%k
n<-dim(IBM_SP)[1]
calc_I22_mean(x, k, p, n, sigma_inv)
}
