% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarScore.R
\name{gene_indep_variable}
\alias{gene_indep_variable}
\title{The matrix of independent variables}
\usage{
gene_indep_variable(samples, p, k)
}
\arguments{
\item{samples}{should be a data frame. Must be greater than three columns, and the first column is a timestamp.}

\item{p}{Lags.}

\item{k}{The dimension of VAR.}
}
\value{
The matrix of independent variables
}
\description{
The matrix of independent variables
}
\examples{
library(VarScore)
gene_indep_variable(IBM_SP, 1, 2)
}
