% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-IBM_SP.R
\docType{data}
\name{IBM_SP}
\alias{IBM_SP}
\title{IBM_SP dataset}
\format{
A data frame with 848 rows and 3 variables:
  \describe{
    \item{date}{date}
    \item{ibm}{numeric}
    \item{sp}{numeric}
  }
}
\source{
Yahoo Finance
}
\usage{
IBM_SP
}
\description{
A dataset containing data on International Business Machines Corporation stock (IBM) and the Standard&Poor’s 500 Index (S&P500) from May 1938 to December 2008.
}
\keyword{datasets}
