% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarScore.R
\name{mean_shift}
\alias{mean_shift}
\title{The Mean-shift perturbation model}
\usage{
mean_shift(B, k, p, samples, sigma, alpha)
}
\arguments{
\item{B}{The hat-B of VAR.}

\item{k}{The dimension of VAR.}

\item{p}{Lags.}

\item{samples}{should be a data frame. Must be greater than three columns, and the first column is a timestamp.}

\item{sigma}{The hat-sigma of VAR.}

\item{alpha}{The significance level.}
}
\value{
The result of Mean-shift perturbation model
}
\description{
The Mean-shift perturbation model
}
\examples{
library(VarScore)
V1<-myVAR(IBM_SP)
new_sigma<-V1$hat_sigma
new_B<-V1$hat_B
residual<-data.frame(t(V1$hat_U))
k<-dim(new_B)[1]
p<-(dim(new_B)[2]-1)\%/\%k

S_mean_shift<-mean_shift(new_B, k, p, samples, new_sigma, 0.05)
ms_scores<-S_mean_shift$all
ms_outliters<-S_mean_shift$outliters
ms_correct<-S_mean_shift$correct
}
