% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarScore.R
\name{Montecarlo}
\alias{Montecarlo}
\title{Montecarlo}
\usage{
Montecarlo(mu, sigma, B, sample_n)
}
\arguments{
\item{mu}{The mean of multivariate normal distribution.}

\item{sigma}{The Σ of multivariate normal distribution.}

\item{B}{The coefficient of VAR.}

\item{sample_n}{The number of simulation.}
}
\value{
data frame
}
\description{
Montecarlo
}
\examples{
library(VarScore)
mu <- matrix(c(0.003,0.006,-0.015),3,1)
sigma <- matrix(c(1,0.6,0.76,0.6,1,0.14,0.76,0.14,1),3,3)
B <- matrix(c(-0.338,-0.134,-2.221,0.746,0.328,4.586,0.058,0.043,0.301),3,3)
}
