% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarScore.R
\name{myVAR}
\alias{myVAR}
\title{The ML estimators of B and Σ}
\usage{
myVAR(samples, p = NULL)
}
\arguments{
\item{samples}{should be a data frame. Must be greater than three columns, and the first column is a timestamp.}

\item{p}{lags. default 'NULL'.}
}
\value{
The list include the ML estimators of B and Σ.
}
\description{
The ML estimators of B and Σ
}
\examples{
library(VarScore)
library(VarScore)
V1<-myVAR(IBM_SP)
new_sigma<-V1$hat_sigma
new_B<-V1$hat_B
residual<-data.frame(t(V1$hat_U))
}
