% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VarScore.R
\name{variance_weight}
\alias{variance_weight}
\title{The Case-weight perturbation model}
\usage{
variance_weight(B, k, p, samples, sigma, alpha)
}
\arguments{
\item{B}{The hat-B of VAR.}

\item{k}{The dimension of VAR.}

\item{p}{Lags.}

\item{samples}{should be a data frame. Must be greater than three columns, and the first column is a timestamp.}

\item{sigma}{The hat-sigma of VAR.}

\item{alpha}{The significance level.}
}
\value{
The result of Case-weight perturbation model
}
\description{
The Case-weight perturbation model
}
\examples{
library(VarScore)
V1<-myVAR(IBM_SP)
new_sigma<-V1$hat_sigma
new_B<-V1$hat_B
residual<-data.frame(t(V1$hat_U))
k<-dim(new_B)[1]
p<-(dim(new_B)[2]-1)\%/\%k

S_variance_weight<-variance_weight(new_B, k, p, IBM_SP, new_sigma, 0.05)
vw_scores<-S_variance_weight$all
vw_outliters<-S_variance_weight$outliters
vw_correct<-S_variance_weight$correct
}
